example/requirements_render.txt
# 渲染功能所需的额外依赖
# 安装命令: pip install -r requirements_render.txt

# 3D渲染引擎
pyrender>=0.1.45
trimesh>=3.9.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 可选：如果遇到OpenGL相关问题，可能需要安装
# pyglet<2  # PyRender在某些系统上需要pyglet 1.x
# PyOpenGL>=3.1.0
# PyOpenGL-accelerate>=3.1.0
